Add simple "minimap" functionailty
This is missing a couple of nice-to-haves, and has some values hardcoded that might be better specified elsewhere, but it works for an MVP.  I started to write the code to convert the HTML to a canvas image myself, but there were too many edge-cases to chase down so I grabbed a library that looks okay.  Had to rearrange the page layout HTML a little, and fix a couple of issues the the ls-corpus HTML too.